Pull to refresh only in accounts part
